Move report emails to separate mailer
changelog: Internal, Emails, Move report emails into separate Mailer
