+++ app/jobs/reports/deleted_user_accounts_report.rb
@@ -20,7 +20,7 @@ module Reports
         issuers = report_hash['issuers']
         report = deleted_user_accounts_data_for_issuers(issuers)
         emails.each do |email|
-          UserMailer.deleted_user_accounts_report(
+          ReportMailer.deleted_user_accounts_report(
             email: email,
             name: name,
             issuers: issuers,
