+++ app/services/db/service_provider_quota_limit/notify_if_any_sp_over_quota_limit.rb
@@ -5,7 +5,7 @@ module Db
         return unless Db::ServiceProviderQuotaLimit::AnySpOverQuotaLimit.call
         email_list = IdentityConfig.store.sps_over_quota_limit_notify_email_list
         email_list.each do |email|
-          UserMailer.sps_over_quota_limit(email).deliver_now_or_later
+          ReportMailer.sps_over_quota_limit(email).deliver_now_or_later
         end
       end
     end
