+++ app/views/report_mailer/deleted_user_accounts_report.html.erb
@@ -1,0 +1,2 @@
+<%= t('report_mailer.deleted_accounts_report.name') %>: <%= @name %><br>
+<%= t('report_mailer.deleted_accounts_report.issuers') %>: <%= @issuers.join(',') %><br>
