--- app/views/user_mailer/deleted_user_accounts_report.html.erb
@@ -1,2 +1,0 @@
-<%= t('user_mailer.deleted_accounts_report.name') %>: <%= @name %><br>
-<%= t('user_mailer.deleted_accounts_report.issuers') %>: <%= @issuers.join(',') %><br>
