+++ config/locales/report_mailer/es.yml
@@ -1,0 +1,11 @@
+---
+es:
+  report_mailer:
+    deleted_accounts_report:
+      issuers: Emisores
+      name: Nombre
+      subject: Informe de cuentas eliminadas
+    sps_over_quota_limit:
+      help: ''
+      info: Hay SP que han superado su límite de cuota.
+      subject: Algunos SP han alcanzado su límite de cuota.
