+++ config/locales/report_mailer/fr.yml
@@ -1,0 +1,11 @@
+---
+fr:
+  report_mailer:
+    deleted_accounts_report:
+      issuers: Émetteurs
+      name: Nom
+      subject: Rapport sur les comptes supprimés
+    sps_over_quota_limit:
+      help: ''
+      info: Certains SP ont dépassé leur limite de quota.
+      subject: Certains SP ont atteint leur limite de quota
