+++ spec/mailers/user_mailer_spec.rb
@@ -433,52 +433,6 @@ describe UserMailer, type: :mailer do
     end
   end
 
-  describe '#sps_over_quota_limit' do
-    let(:mail) { UserMailer.sps_over_quota_limit(email_address.email) }
-
-    it_behaves_like 'a system email'
-
-    it 'sends to the current email' do
-      expect(mail.to).to eq [email_address.email]
-    end
-
-    it 'renders the subject' do
-      expect(mail.subject).to eq t('user_mailer.sps_over_quota_limit.subject')
-    end
-
-    it 'renders the body' do
-      expect(mail.html_part.body).
-        to have_content(strip_tags(t('user_mailer.sps_over_quota_limit.info')))
-    end
-  end
-
-  describe 'deleted_user_accounts_report' do
-    let(:mail) do
-      UserMailer.deleted_user_accounts_report(
-        email: email_address.email,
-        name: 'my name',
-        issuers: %w[issuer1 issuer2],
-        data: 'data',
-      )
-    end
-
-    it_behaves_like 'a system email'
-
-    it 'sends to the current email' do
-      expect(mail.to).to eq [email_address.email]
-    end
-
-    it 'renders the subject' do
-      expect(mail.subject).to eq t('user_mailer.deleted_accounts_report.subject')
-    end
-
-    it 'renders the body' do
-      expect(mail.html_part.body).to have_content('my name')
-      expect(mail.html_part.body).to have_content('issuer1')
-      expect(mail.html_part.body).to have_content('issuer2')
-    end
-  end
-
   describe '#account_verified' do
     disavowal_token = 'i_am_disavowal_token'
     let(:sp_name) { '' }
